requirements.txt
asn1crypto==0.24.0
certifi==2018.11.29
cffi==1.12.2
cryptography==2.6.1
Django==2.1.7
mysqlclient==1.4.2.post1
netaddr==0.7.19
pycparser==2.19
pyOpenSSL==19.0.0
python-dateutil==2.8.0
pytz==2018.9
six==1.12.0
text-unidecode==1.2